+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## 1.1.0
 
+* Core: Add pac file server
+  * Update rules from gfwlist
+  * Support adding user rules
+
+* Core: Seperate log files
+
+* Extra: Add benchmark
+
 ## 1.0.4(2016-05-21)
 
 * Bug fix: typo `clientToDst.resumse()`